revert: remove .float() cast causing OOM across 24 layers

+++ cortexia/block.py
@@ -349,7 +349,7 @@ class SharedCacheTransformerLayer(nn.Module):
         k_r = K_rotate.transpose(1, 2).expand(-1, nh, -1, -1)
         s_r = torch.matmul(q_r, k_r.transpose(-2, -1)) * scale_r
 
-        scores = (s_c + s_r).float()
+        scores = s_c + s_r
         if self.attn_logit_cap is not None:
             scores = torch.tanh(scores / self.attn_logit_cap) * self.attn_logit_cap
 
@@ -403,7 +403,7 @@ class SharedCacheTransformerLayer(nn.Module):
             K_state = self.k_norm(K_state)
 
         scores = self._decoupled_scores(Q_state, Q_rotate, K_state, K_rotate, T)
-        attn_w = F.softmax(scores, dim=-1).to(Q_state.dtype)
+        attn_w = F.softmax(scores, dim=-1)
         attn_w = self.attn_dropout(attn_w)
 
         n = self.num_heads // self.num_kv_groups
@@ -471,7 +471,7 @@ class SharedCacheTransformerLayer(nn.Module):
                 K_state = self.k_norm(K_state)
 
         scores = self._decoupled_scores(Q_state, Q_rotate, K_state, K_rotate, S_new, T, S_new > 1)
-        attn_w = F.softmax(scores, dim=-1).to(Q_state.dtype)
+        attn_w = F.softmax(scores, dim=-1)
         attn_w = self.attn_dropout(attn_w)
 
         n = self.num_heads // self.num_kv_groups
